[FEAT-13] combine form and modal elements to one, refactor accordingly
