+++ src/app/notes/page.tsx
@@ -1,5 +1,4 @@
 
-import Modal from "@/components/Modal";
 import NewTopicFormModal from "@/components/NewTopicFormModal";
 
 export default async function NotesPage() {
@@ -7,7 +6,6 @@ export default async function NotesPage() {
   return (
     <section aria-label="Notes">
       <h1 className="text-4xl font-bold mt-8 mb-4" >Notes</h1>
-      <NewTopicFormModal />
       <p>Select some notes to get started</p>
     </section>
   );
